% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Load dataset from various formats}
\usage{
load_dataset(file_path)
}
\arguments{
\item{file_path}{The path to the data file.}
}
\value{
A data frame containing the loaded data
}
\description{
This function loads datasets from .csv, .xls, .xlsx,
}
\examples{
data <- load_dataset("data.csv")
data <- load_dataset("data.xls")
data <- load_dataset("data.xlsx")
}
